requirement.txt
gevent==23.9.1
greenlet==3.0.1
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.1.0
zope.event==5.0
zope.interface==6.1
